# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: diffx API
  version: 1.0.0
  description: Push-only diff ingestion and review API, shared by local and remote servers.
servers:
  - url: /
paths:
...
  /api/v2/ingestion-jobs:
    post:
      operationId: acceptIngestion
      description: Durably accept an immutable submission. Identical retries return the same job; changed payloads conflict. Acceptance is not publication.
      security:
        - bearerAuth: []
      parameters:
        - name: X-Diffx-State
          in: header
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestionRequest"
      responses:
        "202":
          description: Durable job accepted; poll its Location until succeeded or failed.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestionJob"
        default:
          $ref: "#/components/responses/Problem"
...
  /api/v2/events:
    get:
      operationId: watchIngestions
      description: Server-sent ingestion notifications. Event data contains contextId; query the catalog for durable state after reconnecting.
      responses:
        "200":
          description: Server-sent events named ingestion, with JSON data containing contextId.
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  responses:
    Problem:
      description: Request failed.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Comparison:
      type: object
      description: Immutable Git comparison facts captured by the producer and stored with the observation. Applies to the all scope; staged always compares HEAD to index and unstaged compares index to working tree. Missing comparison metadata retains legacy HEAD semantics.
      additionalProperties: false
      required: [kind, baseRef]
      properties:
        kind:
          type: string
          enum: [working-tree, branch]
          description: Working-tree compares HEAD to working changes. Branch compares the default or explicit base ref's merge base to the checkout, including committed and working changes; object-only branch recovery compares that merge base to HEAD.
        baseRef:
          type: string
          description: Logical comparison baseline. HEAD for working changes; normalized Git ref for branch comparisons.
        baseCommit:
          type: string
          description: Resolved baseline tip at collection time. Absent for unborn checkouts.
        mergeBase:
          type: string
          description: Full commit ID actually used as the all-scope baseline. For branch comparisons this is the merge base of baseCommit and observation.head, which may differ from the base ref tip. For working-tree comparisons it equals the captured HEAD; absent for unborn checkouts.
    AgentSession:
      type: object
      additionalProperties: false
      required: [harness, id]
      properties:
        harness:
          type: string
        id:
          type: string
        name:
          type: string
...
    ObservationMetadata:
      type: object
      additionalProperties: false
      required:
        [
          sourceId,
          hostname,
          runId,
          agent,
          trigger,
          repositoryKey,
          repositoryName,
          remoteUrl,
          checkoutKey,
          root,
          worktreeName,
          branch,
          head,
          collectedAt,
          collectorVersion,
        ]
      properties:
        sourceId:
          type: string
          description: Producer-provided source identity, independent of hostname and path.
        hostname:
          type: string
        runId:
          type: string
        agent:
          type: string
        trigger:
          type: string
        repositoryKey:
          type: string
        repositoryName:
          type: string
        remoteUrl:
          type: string
        checkoutKey:
          type: string
        linkedWorktree:
          type: boolean
          description: True for a linked Git worktree, false for the primary checkout. Omitted when unknown in older snapshots.
        root:
          type: string
          description: Source-local path; the server does not access it.
        worktreeName:
          type: string
        branch:
          type: string
        branchId:
          type: string
          description: Stable producer branch identity, preserved through branch renames. Optional for older producers and detached HEAD.
        head:
          type: [string, "null"]
        collectedAt:
          type: integer
          format: int64
          description: Producer collection time in Unix milliseconds.
        collectorVersion:
          type: string
        comparison:
          $ref: "#/components/schemas/Comparison"
        agentSession:
          $ref: "#/components/schemas/AgentSession"
        triggerRoot:
          type: string
          description: Directory whose manual command or agent event triggered collection; does not establish authorship.
    IngestionScope:
      type: object
      additionalProperties: false
      required: [snapshot, patches]
      properties:
        snapshot:
          $ref: "#/components/schemas/RepositoryDiff"
        patches:
          type: object
          description: File previews keyed by snapshot file ID.
          additionalProperties:
            $ref: "#/components/schemas/FilePatch"
    IngestionJob:
      type: object
      additionalProperties: false
      required: [id, submissionId, state, attempt]
      properties:
        id:
          type: string
        submissionId:
          type: string
        state:
          type: string
          enum: [queued, running, retrying, succeeded, failed]
        attempt:
          type: integer
        contextId:
          type: string
        detail:
          type: string
    IngestionRequest:
      type: object
      additionalProperties: false
      required: [protocolVersion, submissionId, metadata, scopes]
      properties:
        protocolVersion:
          type: integer
          enum: [4]
          description: Current producers send 2. Legacy version 1 remains accepted for ingestion and replay.
        submissionId:
          type: string
          minLength: 1
          description: Retry identity. Reusing it with different content fails.
        contentHash:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Optional SHA256 of complete contents and staging state, independent of preview truncation and timestamps. Omit when complete identity cannot be proven.
        metadata:
          $ref: "#/components/schemas/ObservationMetadata"
        scopes:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: "#/components/schemas/IngestionScope"
...
    DiffMode:
      type: string
      enum: [all, staged, unstaged]
...
    ChangedFile:
      type: object
      additionalProperties: false
      required:
        [
          id,
          path,
          oldPath,
          status,
          indexStatus,
          worktreeStatus,
          additions,
          deletions,
          binary,
          fingerprint,
        ]
      properties:
        id:
          type: string
        path:
          type: string
        oldPath:
          type: [string, "null"]
        status:
          type: string
        indexStatus:
          type: string
        worktreeStatus:
          type: string
        additions:
          type: integer
        deletions:
          type: integer
        binary:
          type: boolean
        fingerprint:
          type: string
        recreated:
          type: boolean
    RepositoryDiff:
      type: object
      additionalProperties: false
      required:
        [
          id,
          versionId,
          locationId,
          repositoryId,
          source,
          root,
          name,
          branch,
          head,
          mode,
          files,
          revision,
        ]
      properties:
        id:
          type: string
        versionId:
          type: string
        locationId:
          type: [string, "null"]
        repositoryId:
          type: [string, "null"]
        source:
          type: string
          enum: [local, stdin]
        root:
          type: string
        name:
          type: string
        branch:
          type: string
        head:
          type: [string, "null"]
        mode:
          $ref: "#/components/schemas/DiffMode"
        files:
          type: array
          items:
            $ref: "#/components/schemas/ChangedFile"
        revision:
          type: string
    FilePatch:
      type: object
      additionalProperties: false
      required: [patch, message]
      properties:
        patch:
          type: string
        message:
          type: [string, "null"]
        contents:
          $ref: "#/components/schemas/FileContents"
    FileContents:
      type: object
      additionalProperties: false
      required: [before, after]
      properties:
        before:
          type: string
        after:
          type: string
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, detail]
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
        capability:
          type: string